Remove dependency on postgresql shell client commands
Remove dependency on PostgresSQL shell wrappers being installed
(createdb, dropdb, pg_dump). Instead use Sequel to issue SQL DDL
commands.

This does require swapping `pg_dump` for Sequel's schema dump tool which
is only capable of dumping the schema in the form of a Sequel data
migration, but this should still satisfy the current goal of dumping
the schema to ensure it is always in a consistent state.

The goal of this is to create a Docker image for this application that
will allow the execution of the migration commands from the docker
image without having to translate the connection options to flags on the
shell commands.